HZC-521 - Use int instead of byte for the iteration counter

+++ src/main/java/com/hazelcast/cloud/Client.java
@@ -32,7 +32,7 @@ public class Client {
 
         IMap<String, String> map = client.getMap("map");
         Random random = new Random();
-        byte iterationCounter = 0;
+        int iterationCounter = 0;
         while (true) {
             int randomKey = random.nextInt(100_000);
             map.put("key-" + randomKey, "value-" + randomKey);
